Show the original footnote numbers and text when the page is printed
The previous behavior was to wrap footnote links and bodies in an inline
"display: none". Now these elements are given the style
"footnote-print-only", which is only displayed while printing.
Conversely the Bigfoot ellipsis button is hidden during printing.

(If "actionOriginalFN" is "delete", then all footnote-related elements
are destroyed, just like before, and won't be displayed during printing.
The ellipsis button is still hidden in this case.)